+++ lib/flutter_image_picker.dart
@@ -4,7 +4,7 @@
 ///
 library flutter_image_picker;
 
-export 'src/models/image_picker_config.dart';
-export 'src/models/image_picker_theme.dart';
-export 'src/services/image_picker_service.dart';
-export 'src/ui/image_picker.dart';
+export "src/models/image_picker_config.dart";
+export "src/models/image_picker_theme.dart";
+export "src/services/image_picker_service.dart";
+export "src/ui/image_picker.dart";